add entropic module
