+++ matgen/base.py
@@ -17,6 +17,7 @@ from matplotlib.axes import Axes
 from mpl_toolkits.mplot3d.art3d import Poly3DCollection
 
 from matgen import matutils
+from matgen.entropic import TripleJunctionSet
 
 
 class Cell:
@@ -1039,164 +1040,6 @@ class Poly(Grain):
     @property
     def gb_ids(self):
         return self.f_ids
-
-
-class TripleJunctionSet:
-    """
-    """
-
-    def __init__(self, p, j_tuple) -> None:
-        """
-        """
-        self.p = p
-        self.j0, self.j1, self.j2, self.j3 = j_tuple
-
-    @property
-    def q(self):
-        """
-        """
-        return 1 - self.p
-
-    @property
-    def j_tuple(self):
-        """
-        """
-        return (self.j0, self.j1, self.j2, self.j3)
-
-    @property
-    def Sp(self):
-        """
-        """
-        if self.p == 0 or self.p == 1:
-            return 0
-        else:
-            return matutils.entropy(self.p)
-
-    @property
-    def Sp_m(self):
-        """
-        """
-        if self.p == 0:
-            return np.inf
-        else:
-            return matutils.entropy_m(self.p)
-
-    @property
-    def Sp_s(self):
-        """
-        """
-        if self.p == 0:
-            return - np.inf
-        else:
-            return matutils.entropy_s(self.p)
-
-    @property
-    def p_expected(self):
-        """
-        """
-        return (self.j1 + 2*self.j2 + 3*self.j3) / 3
-        
-    @property
-    def delta_p(self):
-        """
-        """
-        return abs(self.p_expected - self.p)
-
-    @property
-    def S(self):
-        """
-        """
-        return matutils.entropy(*self.j_tuple)
-
-    @property
-    def S_m(self):
-        """
-        """
-        if np.any(np.array(self.j_tuple) == 0):
-            return np.inf
-        else:
-            return matutils.entropy_m(*self.j_tuple)
-
-    @property
-    def S_s(self):
-        """
-        """
-        if np.any(np.array(self.j_tuple) == 0):
-            return - np.inf
-        else:
-            return matutils.entropy_s(*self.j_tuple)
-
-    @property
-    def kappa(self):
-        """
-        """
-        if self.S_s == 0:
-            return 0
-        else:
-            return self.S_m / self.S_s
-
-    @property
-    def delta_S(self):
-        """
-        """
-        return self.Sp - self.S
-
-    @property
-    def d_tuple(self):
-        """
-        """
-        return matutils.get_d_tuple(self.j_tuple)
-
-    @property
-    def d1(self):
-        """
-        """
-        return self.d_tuple[0]
-
-    @property
-    def d2(self):
-        """
-        """
-        return self.d_tuple[1]
-
-    @property
-    def d3(self):
-        """
-        """
-        return self.d_tuple[2]
-
-    def get_property(self, attr):
-        """
-        """
-        return getattr(self, attr)
-    
-    def get_properties(self, attr_list: list = []) -> Dict:
-        """
-        """
-        if not attr_list:
-            attr_list = [
-                'p',
-                'q',
-                'Sp',
-                'Sp_m',
-                'Sp_s',
-                'S',
-                'S_m',
-                'S_s',
-                'kappa',
-                'delta_S',
-                'd1',
-                'd2',
-                'd3'
-            ]
-
-        try:
-            return {attr_name: getattr(self, attr_name) for attr_name in attr_list}
-        except:
-            logging.exception('Check properties!')
-
-        # values = [getattr(self, attr_name) for attr_name in attr_list]
-        # return pd.DataFrame([values], columns = attr_list)
 
 
 def _add_neighbors(_cells: Dict, _incident_cells: Dict):
